# 2026-cohort/quantum-solvers-equitable-mentorship-pairing-engine/src/openapi.yaml
openapi: 3.1.0
info:
  title: Braid API
  version: 0.1.0
  description: |
    Thin vertical slice: program lookup, application intake, magic-link auth,
    roster, matching runs with a fairness summary, and strands with messages.

    The published form schema is here too, so the apply screen renders what a
    coordinator built rather than a hardcoded set of questions.

    Deliberately excluded until the step that needs them: file upload, and the
    pair-level run actions (swap, lock, reject, re-run).
servers:
  - url: http://localhost:8000/v1
    description: Local FastAPI
  - url: https://api.braid.app/v1
    description: Production
security:
  - sessionCookie: []
...
paths:
...
  /programs/{programId}/strands:
    parameters:
      - $ref: "#/components/parameters/ProgramId"
    get:
      operationId: listStrands
      tags: [Strands]
      summary: List the strands the current account holds in a program
      parameters:
        - name: state
          in: query
          schema:
            $ref: "#/components/schemas/StrandState"
      responses:
        "200":
          description: The strands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StrandSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createStrand
      tags: [Unmatched]
      summary: Pair two people by hand
      description: |
        The remedy for most unmatched reasons. Stored with origin mode `manual`
        so reports can answer whether the algorithm did better than the
        coordinator's hand-picks, which is the whole reason origin mode exists.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrandCreate"
      responses:
        "201":
          description: The strand, active immediately
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strand"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: One of the two already holds a strand, or the mentor is full
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
  /strands/{strandId}:
    parameters:
      - $ref: "#/components/parameters/StrandId"
    get:
      operationId: getStrand
      tags: [Strands]
      summary: Read one strand
      responses:
        "200":
          description: The strand
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strand"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /strands/{strandId}/messages:
    parameters:
      - $ref: "#/components/parameters/StrandId"
    get:
      operationId: listMessages
      tags: [Strands]
      summary: Read a strand's message thread
      description: >
        Newest last. Page backwards through history with `before`, or poll
        forwards for new arrivals with `since`. The two describe opposite
        directions and cannot be combined.
      parameters:
        - name: before
          in: query
          description: Return messages sent before this message id.
          schema:
            type: string
            format: uuid
        - name: since
          in: query
          description: >
            Return only messages sent after this message id, oldest first. A
            thread with nothing new answers with an empty list, which is what
            makes polling cheap.
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: A page of messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagePage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: sendMessage
      tags: [Strands]
      summary: Send a message to a strand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageCreate"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

        "429":
          $ref: "#/components/responses/TooManyRequests"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: braid_session
  parameters:
...
    ProgramId:
      name: programId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    StrandId:
      name: strandId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: The request failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: No valid session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Signed in, but not permitted here
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: No such resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
    TooManyRequests:
      description: Rate limited
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Problem:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable identifier, e.g. `applications_closed`.
        message:
          type: string
          description: Sentence-case text safe to show a person.
        field:
          type: string
          description: Present when the failure belongs to one input.
    Role:
      type: string
      enum: [mentor, mentee]
...
    StrandState:
      type: string
      enum: [draft, active, paused, ended, discarded]
    OriginMode:
      type: string
      enum: [manual, self, batch]
...
    StrandCreate:
      type: object
      required: [menteeParticipationId, mentorParticipationId]
      properties:
        menteeParticipationId:
          type: string
          format: uuid
        mentorParticipationId:
          type: string
          format: uuid
...
    PersonRef:
      type: object
      required: [participationId, name]
      properties:
        participationId:
          type: string
          format: uuid
        name:
          type: string
        photoUrl:
          type: [string, "null"]
          format: uri
...
    StrandMember:
      type: object
      required: [participationId, name, role]
      properties:
        participationId:
          type: string
          format: uuid
        name:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        headline:
          type: [string, "null"]
        photoUrl:
          type: [string, "null"]
          format: uri
        timezone:
          type: [string, "null"]
        skills:
          type: array
          items:
            type: string
    MessagePreview:
      type: object
      required: [authorName, body, sentAt]
      properties:
        authorName:
          type: string
        body:
          type: string
        sentAt:
          type: string
          format: date-time
    StrandSummary:
      type: object
      required: [id, programId, state, originMode, members, unreadCount]
      properties:
        id:
          type: string
          format: uuid
        programId:
          type: string
          format: uuid
        state:
          $ref: "#/components/schemas/StrandState"
        originMode:
          $ref: "#/components/schemas/OriginMode"
        members:
          type: array
          description: Everyone except the current account.
          items:
            $ref: "#/components/schemas/StrandMember"
        lastMessage:
          oneOf:
            - $ref: "#/components/schemas/MessagePreview"
            - type: "null"
        lastActivityAt:
          type: [string, "null"]
          format: date-time
          description: |
            The UI derives the quiet state from this. There is no health enum,
            because a quiet strand is not an error and must not render as one.
        unreadCount:
          type: integer
        nextSessionAt:
          type: [string, "null"]
          format: date-time
        endedAt:
          type: [string, "null"]
          format: date-time
          description: |
            Set when `state` is `ended`. The card names the date, so the
            summary has to carry it rather than making the list fetch each
            strand to find out when it finished.
    Strand:
      type: object
      required: [id, programId, state, originMode, members, createdAt]
      properties:
        id:
          type: string
          format: uuid
        programId:
          type: string
          format: uuid
        state:
          $ref: "#/components/schemas/StrandState"
        originMode:
          $ref: "#/components/schemas/OriginMode"
        members:
          type: array
          description: Every member, including the current account.
          items:
            $ref: "#/components/schemas/StrandMember"
        matchRationale:
          type: [string, "null"]
          description: |
            Why these people were put together. A partial explanation, not the
            explanation: in a global assignment the real reason often involves a
            third person. Present it as such.
        createdAt:
          type: string
          format: date-time
        endedAt:
          type: [string, "null"]
          format: date-time
    MessageCreate:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 10000
        clientToken:
          type: string
          description: |
            Echoed back on the created message so an optimistic send can be
            reconciled against the server row instead of duplicated.
    Message:
      type: object
      required: [id, strandId, author, body, sentAt, deliveryState]
      properties:
        id:
          type: string
          format: uuid
        strandId:
          type: string
          format: uuid
        author:
          $ref: "#/components/schemas/PersonRef"
        body:
          type: string
        sentAt:
          type: string
          format: date-time
        deliveryState:
          type: string
          enum: [sent, delivered]
        clientToken:
          type: [string, "null"]
    MessagePage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        nextCursor:
          type: [string, "null"]
          description: Pass as `before` to fetch the previous page. Null when at the start.